+++ ese_motors_encoders/Core/Src/BOTterfly-C/encoder.c
@@ -36,7 +36,7 @@ uint8_t ENC_InitTimer(ENC_HandleTypeDef* Encoder, TIM_HandleTypeDef *htim, uint3
  *	@param	Encoder is a ENC_HandleTypeDef
  *	@param 	ticksPerRev value
  */
-void ENC_SetTicksPerRev(ENC_HandleTypeDef* Encoder, uint16_t ticksPerRev){
+void ENC_SetTicksPerRev(ENC_HandleTypeDef* Encoder, float ticksPerRev){
 	Encoder->TicksPerRev = ticksPerRev;
 }
